more field cleanup
