+++ src/f2e16.rs
@@ -1,9 +1,3 @@
-use derive_more::{Add, AddAssign, BitXor, BitXorAssign, Sub, SubAssign};
-
-// must be placed in a separate file, such that the preproc never tries to eval OUT_DIR
-// in env which does not exist in the build.rs case
-#[cfg(table_bootstrap_complete)]
-include!(concat!(env!("OUT_DIR"), "/table_f2e16.rs"));
 
 pub type Elt = u16;
 pub type Wide = u32;
@@ -11,93 +5,13 @@ pub type Wide = u32;
 pub const FIELD_BITS: usize = 16;
 pub const FIELD_SIZE: usize = 1_usize << FIELD_BITS;
 
+include!("f_log_mul.rs");
 
-/// Additive via XOR form of f2e16
-#[derive(Clone, Copy, Debug, Default, BitXor, BitXorAssign, PartialEq, Eq)] // PartialOrd,Ord
-pub struct Additive(pub Elt);
 
-impl Additive {
-    #[inline(always)]
-	pub fn to_wide(self) -> Wide {
-		self.0 as Wide
-	}
-    #[inline(always)]
-	pub fn from_wide(x: Wide) -> Additive {
-		Additive(x as Elt)
-	}
-
-	pub const ZERO: Additive = Additive(0u16);
-	// pub const ONE: Additive = Additive(???);
-}
-
+// must be placed in a separate file, such that the preproc never tries to eval OUT_DIR
+// in env which does not exist in the build.rs case
 #[cfg(table_bootstrap_complete)]
-impl Additive {
-	/// Return multiplier prepared form
-    #[inline(always)]
-	pub fn to_multiplier(self) -> Multiplier {
-		Multiplier(LOG_TABLE[self.0 as usize])
-	}
-
-	/// Return a*EXP_TABLE[b] over GF(2^r)
-    #[inline(always)]
-	pub fn mul(self, other: Multiplier) -> Additive {
-		if self == Self::ZERO {
-			return Self::ZERO;
-		}
-		let log = (LOG_TABLE[self.0 as usize] as u32) + other.0 as u32;
-		let offset = (log & ONEMASK as u32) + (log >> FIELD_BITS);
-		Additive(EXP_TABLE[offset as usize])
-	}
-
-	/// Multiply field elements by a single multiplier, using SIMD if available
-    #[inline(always)]
-	pub fn mul_assign_slice(selfy: &mut [Self], other: Multiplier) {
-		// TODO: SIMD
-		for s in selfy {
-			*s = s.mul(other);
-		}
-	}
-}
-
-
-/// Multiplicaiton friendly LOG form of f2e16
-#[derive(Clone, Copy, Debug, Add, AddAssign, Sub, SubAssign, PartialEq, Eq)] // Default, PartialOrd,Ord
-pub struct Multiplier(pub u16);
-
-impl Multiplier {
-    #[inline(always)]
-	pub fn to_wide(self) -> u32 {
-		self.0 as u32
-	}
-    #[inline(always)]
-	pub fn from_wide(x: u32) -> Multiplier {
-		Multiplier(x as u16)
-	}
-}
-
-
-/// Fast Walsh–Hadamard transform over modulo ONEMASK
-pub fn walsh(data: &mut [Multiplier], size: usize) {
-	let mut depart_no = 1_usize;
-	while depart_no < size {
-		let mut j = 0;
-		let depart_no_next = depart_no << 1;
-		while j < size {
-			for i in j..(depart_no + j) {
-				// We deal with data in log form here, but field form looks like:
-				//			 data[i] := data[i] / data[i+depart_no]
-				// data[i+depart_no] := data[i] * data[i+depart_no]
-				let mask = ONEMASK as Wide;
-				let tmp2: Wide = data[i].to_wide() + mask - data[i + depart_no].to_wide();
-				let tmp1: Wide = data[i].to_wide() + data[i + depart_no].to_wide();
-				data[i] = Multiplier(((tmp1 & mask) + (tmp1 >> FIELD_BITS)) as Elt);
-				data[i + depart_no] = Multiplier(((tmp2 & mask) + (tmp2 >> FIELD_BITS)) as Elt);
-			}
-			j += depart_no_next;
-		}
-		depart_no = depart_no_next;
-	}
-}
+include!(concat!(env!("OUT_DIR"), "/table_f2e16.rs"));
 
 
 /* Needs Cleanup  */
@@ -107,6 +21,10 @@ pub const ONEMASK: Elt = (FIELD_SIZE - 1) as Elt;
 
 /// Quotient ideal generator given by tail of irreducible polynomial
 pub const GENERATOR: Elt = 0x2D; // x^16 + x^5 + x^3 + x^2 + 1
+
+// impl Additive {
+//     pub const ONE: Additive = Additive(???);
+// }
 
 // Cantor basis
 pub const BASE: [Elt; FIELD_BITS] =
